Run eslint on App.jsx and fix all errors
